Add files via upload
pdf to excel script that will read a pdf file and pull the items information as long as it follows a certain regex. Then creates a new excel sheet with all the items compiled in one sheet and also another file using a template file provided. Currently 20% accurate can be imporoved upon.